refactor(cognition): SettleOutcome + SettleStep extracted from act_observe (pure code-motion)
Moved the two settlement wire types — SettleOutcome (struct + infra_failure impl) and
SettleStep (enum + from_settled impl) — verbatim into a new types.rs, re-exported via
`pub use types::{SettleOutcome, SettleStep}` so no external caller changes. Behavior-identical:
only imports/visibility/re-exports touched, no logic. mod.rs 2899 → 2778 lines; types.rs is 135
lines. act_observe tests green: 28 passed, 0 failed.
