+++ core/continuum-core/src/cognition/act_observe/mod.rs
@@ -44,68 +44,12 @@ use perception::{
     all_calls_already_satisfied, collect_touched_paths, is_redundant_orientation,
 };
 
+mod types;
+pub use types::{SettleOutcome, SettleStep};
+
 // The working-memory trail-head bound lives in `working_memory.rs` now (its home — WM owns
 // its own truncation). Still used here for the settlement answer-head.
 
-/// The result of driving a mind to settlement.
-pub struct SettleOutcome {
-    /// The verdict the mind settled on: `Speak`/`RaiseUnprompted`/`Pass` when it
-    /// settled, or the final un-driven `Act` if the external budget ran out
-    /// mid-action (the grader grades that as "did not finish" — honest, never a
-    /// fabricated answer).
-    pub decision: Decision,
-    /// The spoken text, present only when the settled decision is `Speak` /
-    /// `RaiseUnprompted`. This is what an external observer (the grader, or a peer
-    /// in the room) reads.
-    pub spoken: Option<String>,
-    /// How many actions were executed before settling.
-    pub acts: usize,
-    /// The final world-state, with each action's observation folded in — what the
-    /// last tick perceived. Captured for replay/forensics.
-    pub world_state: String,
-    /// The accumulated cost of settling this task: every act→observe deliberation
-    /// generation's latency + tokens, summed. `tokens_per_second()` re-derives
-    /// throughput from the totals. This is the speed/latency the eval reports next
-    /// to the accuracy grade — the same number a live turn could surface for the
-    /// serving governor.
-    pub metrics: TurnMetrics,
-    /// `Some(cause)` when the settle loop stopped because the deliberation model
-    /// call FAILED (timeout, 5xx, a serving lane refusing a model it isn't hosting)
-    /// rather than settling on a verdict. The grader MUST treat this as an
-    /// infrastructure failure — NOT a wrong answer — so an inference hiccup never
-    /// corrupts the accuracy metric ([[self-improvement-is-a-control-loop]]: the
-    /// reward is only as trustworthy as the metric). `None` on every settled turn.
-    pub inference_error: Option<String>,
-    /// File paths her acts NAMED (any `file_path`/`path` string arg on an executed
-    /// call), in first-touch order, deduped. This is the turn's investigation trail
-    /// as STATE — distinct from `files_changed` (what git saw): a failed edit or a
-    /// read appears here and nowhere else. Exists because an N-chances retry is a
-    /// FRESH turn with fresh working memory (glass-boxed 2026-08-08,
-    /// benchy-sympy-22840-n4 attempt 2: the verdict said "the file you already
-    /// identified" and she had no memory of identifying it — 10 acts re-deriving
-    /// cse_main.py). The retry caller threads these into the next attempt's task.
-    pub touched_paths: Vec<String>,
-}
-
-impl SettleOutcome {
-    /// An infra-failure outcome: no verdict was reached because the deliberation
-    /// call failed OR was aborted by a watchdog (e.g. a per-task deadline in the
-    /// eval loop). The grader keys on `inference_error.is_some()` to score this a
-    /// NAMED infrastructure failure — never a wrong answer — so a serving wedge
-    /// never masquerades as a capability miss ([[self-improvement-is-a-control-loop]]).
-    /// Zeroed metrics/acts because none accrued meaningfully. `TurnMetrics: Default`.
-    pub fn infra_failure(cause: impl Into<String>) -> Self {
-        Self {
-            decision: Decision::Pass,
-            spoken: None,
-            acts: 0,
-            world_state: String::new(),
-            metrics: TurnMetrics::default(),
-            inference_error: Some(cause.into()),
-            touched_paths: Vec::new(),
-        }
-    }
-}
 
 
 /// Recall salience for an action-observation receipt (#166). Below the neutral
@@ -882,71 +826,6 @@ pub async fn drive_to_settle(
     }
 }
 
-/// The outcome of ONE settlement [`settle_step`].
-#[derive(Debug)]
-pub enum SettleStep {
-    /// She settled on speech (`Speak`/`RaiseUnprompted`) — the prose turn an
-    /// observer (a peer, or the grader) reads.
-    Spoke(String),
-    /// She reached for her hands AND the act was carried out; the result is admitted
-    /// as memory + a stamped proprioception trace. The caller re-perceives next
-    /// (live: next metronome tick; eval: next loop step). The calls+intent ride
-    /// along so a caller that paces acting (the eval budget) can report the final
-    /// Act if its budget runs out on the following step.
-    Acted { calls: Vec<ToolCall>, intent: String },
-    /// She decided to act but the caller's budget said no this step (`may_act =
-    /// false`) — the act was NOT executed. Only the eval driver passes `may_act =
-    /// false`; the live heartbeat always permits its one act, so it never sees this.
-    WouldAct { calls: Vec<ToolCall>, intent: String },
-    /// She chose silence (`Pass`) — honored as a turn that produces no utterance.
-    Passed,
-    /// She reached for an act that could NOT be carried out (no hands / executor
-    /// error). No utterance; the intent rides along for honest logging/grading.
-    ActUnfulfilled { calls: Vec<ToolCall>, intent: String },
-    /// The deliberation model call itself FAILED — a timeout, a 5xx, or the serving
-    /// lane refusing a model it isn't hosting (the swept-model bug). NO verdict was
-    /// produced. This is NOT a `Passed`: a failed model is not a chosen silence
-    /// ([[fallbacks-are-illegal-fail-loud]]). Every caller surfaces the cause LOUD —
-    /// the command reports `inferenceFailed`, the live heartbeat logs + retries next
-    /// tick, the eval grades it an infra failure — never a serene no-op that hides a
-    /// broken lane. `error` names the cause verbatim from the adapter.
-    InferenceFailed { error: String },
-}
-
-impl SettleStep {
-    /// Project a fully-driven [`SettleOutcome`] back onto the SAME `SettleStep` the
-    /// live heartbeat already handles — so a DIRECTED live turn can `drive_to_settle`
-    /// (converge to an answer in-turn, exactly as the eval path does) and feed the
-    /// result through the one existing turn handler, no parallel match.
-    ///
-    /// The mapping is total and lossless for the live handler's purposes:
-    ///   • `inference_error` present → `InferenceFailed` (a failed model is never a
-    ///     chosen silence — [[fallbacks-are-illegal-fail-loud]]);
-    ///   • `Speak`/`RaiseUnprompted` → `Spoke` (the prose reaches the room);
-    ///   • `Act` → `Acted` — the drive spent its whole act budget without settling on
-    ///     speech; the results are already in memory, so the live handler `continue`s
-    ///     and the metronome re-perceives next tick (the honest long-tail degrade, and
-    ///     the ONLY case a directed turn still leans on the tick loop);
-    ///   • `Pass` → `Passed` (she declined in her own words).
-    /// `drive_to_settle` collapses `WouldAct`/`ActUnfulfilled` into `Decision::Act`
-    /// before returning, so those never surface here — an over-budget or un-carried
-    /// act both land on `Acted`, which the live handler treats as "re-perceive next
-    /// tick", the correct move either way.
-    pub fn from_settled(outcome: SettleOutcome) -> (SettleStep, Option<TurnMetrics>) {
-        let metrics = Some(outcome.metrics);
-        if let Some(error) = outcome.inference_error {
-            return (SettleStep::InferenceFailed { error }, metrics);
-        }
-        let step = match outcome.decision {
-            Decision::Speak { text } | Decision::RaiseUnprompted { text } => {
-                SettleStep::Spoke(text)
-            }
-            Decision::Act { calls, intent } => SettleStep::Acted { calls, intent },
-            Decision::Pass => SettleStep::Passed,
-        };
-        (step, metrics)
-    }
-}
 
 /// True only for a REAL receipt line — `[action #<digit>`. The proprioception
 /// TEACHING texts mention the literal placeholder `[action #n]` ("real
